refactor: reorganise code and remove things we don't need
