+++ app/core/news/template/digest.njk.jinja
